Initial commit of working model
Create framework for the model part of the UI. This includes support for
different modes (currently only pairing and normal mode exist). The
model is threaded to run in parallel to the display app, and it has a
flag that can be set to signal a stop. The only feature that has been
linked to the model so far is the pairing mode button